+++ R/ir_parse.R
@@ -19,6 +19,13 @@ ir_parse <- function(x, options, type = NULL) {
 
   eqs <- ir_parse_arrays(eqs, variables, config$include$names, source)
 
+  ## This performs a round of optimisation where we try to simplify
+  ## away expressions for the dimensions, which reduces the number of
+  ## required variables.
+  if (options$rewrite_dims && features$has_array) {
+    eqs <- ir_parse_rewrite_dims(eqs)
+  }
+
   packing <- ir_parse_packing(eqs, variables, source)
   eqs <- c(eqs, packing$offsets)
   packing$offsets <- NULL
@@ -318,13 +325,18 @@ ir_parse_stage <- function(eqs, dependencies, variables, time_name, source) {
   stage[names_if(vlapply(eqs, is_null))] <- STAGE_NULL
 
   i <- vlapply(eqs, function(x) !is.null(x$array))
-  len <- unique(vcapply(eqs[i], function(x) x$array$dimnames$length))
-  err <- stage[len] == STAGE_TIME
+  len <- lapply(eqs[i], function(x) x$array$dimnames$length)
+  ## We end up with sometimes a string and sometimes a symbol here
+  ## which is unsatisfactory.
+  len_var <- vcapply(len[!vlapply(len, is.numeric)], as.character)
+  err <- stage[len_var] == STAGE_TIME
 
   if (any(err)) {
+    ## TODO: in the case where we rewrite dimensions this error is not
+    ## great beause we've lost the dim() call!
     ir_parse_error(
       "Array extent is determined by time",
-      ir_parse_error_lines(eqs[len[err]]), source)
+      ir_parse_error_lines(eqs[len_var[err]]), source)
   }
 
   stage
@@ -342,8 +354,11 @@ ir_parse_packing_new <- function(eqs, variables, offset_prefix) {
 
 ir_parse_packing_internal <- function(names, rank, len, variables,
                                       offset_prefix) {
-  ## We'll pack from least to most complex:
-  i <- order(rank)
+  ## We'll pack from least to most complex and everything with a fixed
+  ## offset first. This puts all scalars first, then all arrays that
+  ## have compile-time size next (in order of rank), then all arrays
+  ## with user-time size (in order of rank).
+  i <- order(!vlapply(len, is.numeric), rank)
   names <- names[i]
   rank <- rank[i]
   len <- len[i]
@@ -355,10 +370,9 @@ ir_parse_packing_internal <- function(names, rank, len, variables,
   for (i in seq_along(names)) {
     if (!is_array[[i]]) {
       offset[[i + 1L]] <- i
-    } else if (identical(offset[[i]], 0L)) {
-      offset[[i + 1L]] <- as.name(len[[i]])
     } else {
-      offset[[i + 1L]] <- call("+", offset[[i]], as.name(len[[i]]))
+      len_i <- if (is.numeric(len[[i]])) len[[i]] else as.name(len[[i]])
+      offset[[i + 1L]] <- static_eval(call("+", offset[[i]], len_i))
     }
   }
 
@@ -1190,10 +1204,11 @@ ir_parse_delay <- function(eqs, discrete, variables, source) {
       ## TODO: ideally we'd get the correct lines here for source, but
       ## that's low down the list of needs.
       f <- function(x) {
+        depends <- if (is.numeric(x$dim)) character(0) else as.character(x$dim)
         list(name = x$to,
              type = "alloc",
              source = integer(0),
-             depends = ir_parse_depends(variables = x$dim),
+             depends = ir_parse_depends(variables = depends),
              lhs = list(name_data = x$to,
                         name_lhs = x$to,
                         name_equation = x$to),
@@ -1222,8 +1237,11 @@ ir_parse_delay_discrete <- function(eq, eqs, source) {
   nm <- eq$name
   nm_ring <- sprintf("delay_ring_%s", nm)
 
-  depends_ring <- list(functions = character(0),
-                       variables = eq$array$dimnames$length %||% character(0))
+  len <- eq$array$dimnames$length
+  depends_ring <- list(
+    functions = character(0),
+    variables = if (is.character(len)) len else character(0))
+
   lhs_ring <- list(name_data = nm_ring, name_equation = nm_ring,
                    name_lhs = nm_ring, storage_type = "ring_buffer")
   eq_ring <- list(
@@ -1279,19 +1297,27 @@ ir_parse_delay_continuous <- function(eq, eqs, variables, source) {
     substitutions <- list()
   }
 
-  eq_len <- list(
-    name = nm_dim,
-    type = "expression_scalar",
-    source = eq$source,
-    depends = find_symbols(graph$packing$length),
-    lhs = list(name_data = nm_dim, name_equation = nm_dim, name_lhs = nm_dim,
-               storage_type = "int"),
-    rhs = list(value = graph$packing$length))
+  if (is.numeric(graph$packing$length)) {
+    eq_len <- NULL
+    val_len <- graph$packing$length
+    dep_len <- character(0)
+  } else {
+    eq_len <- list(
+      name = nm_dim,
+      type = "expression_scalar",
+      source = eq$source,
+      depends = find_symbols(graph$packing$length),
+      lhs = list(name_data = nm_dim, name_equation = nm_dim, name_lhs = nm_dim,
+                 storage_type = "int"),
+      rhs = list(value = graph$packing$length))
+    val_len <- nm_dim
+    dep_len <- nm_dim
+  }
 
   lhs_use <- eq$lhs[c("name_data", "name_equation", "name_lhs", "special")]
   subs_from <- vcapply(substitutions, "[[", "to")
   depends_use <- join_deps(list(
-    eq$depends, ir_parse_depends(variables = c(nm_dim, subs_from, TIME))))
+    eq$depends, ir_parse_depends(variables = c(dep_len, subs_from, TIME))))
 
   eq_use <- list(
     name = nm,
@@ -1304,7 +1330,7 @@ ir_parse_delay_continuous <- function(eq, eqs, variables, source) {
       state = nm_state,
       index = nm_index,
       substitutions = substitutions,
-      variables = list(length = eq_len$name,
+      variables = list(length = val_len,
                        contents = graph$packing$contents),
       equations = graph$equations,
       default = eq$delay$default,
@@ -1312,7 +1338,7 @@ ir_parse_delay_continuous <- function(eq, eqs, variables, source) {
       depends = eq$delay$depends),
     array = eq$array)
 
-  array <- list(dimnames = list(length = nm_dim, dim = NULL, mult = NULL),
+  array <- list(dimnames = list(length = val_len, dim = NULL, mult = NULL),
                 rank = 1L)
   lhs_index <-
     list(name_data = nm_index, name_equation = nm_index, name_lhs = nm_index,
@@ -1322,7 +1348,7 @@ ir_parse_delay_continuous <- function(eq, eqs, variables, source) {
   offsets <- lapply(variables$contents[match(graph$variables, variable_names)],
                     "[[", "offset")
   depends_index <- join_deps(lapply(offsets, find_symbols))
-  depends_index$variables <- union(depends_index$variables, nm_dim)
+  depends_index$variables <- union(depends_index$variables, dep_len)
   eq_index <- list(
     name = nm_index,
     type = "delay_index",
@@ -1339,7 +1365,7 @@ ir_parse_delay_continuous <- function(eq, eqs, variables, source) {
     name = nm_state,
     type = "null",
     source = eq$source,
-    depends = ir_parse_depends(variables = nm_dim),
+    depends = ir_parse_depends(variables = dep_len),
     lhs = lhs_state,
     array = array)
 
@@ -1348,7 +1374,9 @@ ir_parse_delay_continuous <- function(eq, eqs, variables, source) {
     eq_index$depends$variables <- c(eq_index$depends$variables, names(offsets))
   }
 
-  extra <- c(list(eq_len, eq_index, eq_state, eq_use), offsets)
+  extra <- c(if (is.null(eq_len)) NULL else list(eq_len),
+             list(eq_index, eq_state, eq_use),
+             offsets)
   names(extra) <- vcapply(extra, "[[", "name")
 
   stopifnot(sum(names(eqs) == eq$name) == 1)
@@ -1489,4 +1517,92 @@ ir_parse_expr_rhs_check_inplace <- function(lhs, rhs, line, source) {
       "Expected an array on the lhs of inplace function '%s'", fn),
       line, source)
   }
+}
+
+
+## This approach could probably be applied over the whole tree really,
+## as we might be able to eliminate some other compile time
+## things. However, doing that will make the models less debuggable.
+ir_parse_rewrite_dims <- function(eqs) {
+  compute <- function(x) {
+    if (is.numeric(x)) {
+      x
+    } else if (is.symbol(x)) {
+      x_eq <- eqs[[deparse_str(x)]]
+      ## use identical() here to cope with x_eq being NULL when 't' is
+      ## passed through (that will be an error elsewhere).
+      if (identical(x_eq$type, "expression_scalar")) {
+        compute(x_eq$rhs$value)
+      } else {
+        x
+      }
+    } else if (is_call(x, "length")) {
+      ## NOTE: use array_dim_name because we might hit things like
+      ## length(y) where 'y' is one of the variables; we can't look up
+      ## eqs[[name]]$array$length without checking that.
+      compute(as.name(array_dim_name(as.character(x[[2]]))))
+    } else if (is_call(x, "dim")) {
+      compute(as.name(array_dim_name(as.character(x[[2]]), x[[3]])))
+    } else if (is.recursive(x)) {
+      x[-1] <- lapply(x[-1], compute)
+      x
+    } else { # NULL
+      x
+    }
+  }
+
+  ## alternatively look in all $array$dimnames elements
+  nms <- grep("dim_", names(eqs), value = TRUE)
+
+  val <- lapply(eqs[nms], function(eq)
+    static_eval(compute(eq$rhs$value)))
+
+  rewrite <- vlapply(val, function(x) is.symbol(x) || is.numeric(x))
+
+  subs <- val[rewrite]
+
+  ## Try and deduplicate the rest. However, it's not totally obvious
+  ## that we can do this without creating some weird dependency
+  ## issues.
+  leave <- val[!rewrite]
+  ## Do not deduplicate NULL dimensions; these are set by user() later.
+  dup <- duplicated(leave) & !vlapply(leave, is.null)
+  if (any(dup)) {
+    i <- match(leave[dup], leave)
+    subs <- c(subs,
+              set_names(lapply(names(leave)[i], as.name), names(leave)[dup]))
+  }
+
+  replace <- function(x, y) {
+    i <- match(vcapply(x, function(x) x %||% ""), names(y))
+    j <- which(!is.na(i))
+    x[j] <- unname(y)[i[j]]
+    na_drop(x)
+  }
+
+  subs_env <- list2env(subs, parent = emptyenv())
+  subs_dep <- vcapply(subs, function(x)
+    if (is.numeric(x)) NA_character_ else deparse_str(x))
+
+  rewrite_eq <- function(eq) {
+    eq$rhs$value <- substitute_(eq$rhs$value, subs_env)
+
+    eq$depends$variables <- replace(eq$depends$variables, subs_dep)
+    eq$lhs$depends$variables <- replace(eq$lhs$depends$variables, subs_dep)
+
+    if (!is.null(eq$array$dimnames)) {
+      eq$array$dimnames$length <- replace(eq$array$dimnames$length, subs)[[1]]
+      eq$array$dimnames$dim <- replace(eq$array$dimnames$dim, subs)
+      eq$array$dimnames$mult <- replace(eq$array$dimnames$mult, subs)
+    }
+
+    if (!is.null(eq$delay)) {
+      eq$delay$depends$variables <-
+        replace(eq$delay$depends$variables, subs_dep)
+    }
+
+    eq
+  }
+
+  lapply(eqs[setdiff(names(eqs), names(subs))], rewrite_eq)
 }