+++ tests/testthat/helper-odin.R
@@ -128,8 +128,9 @@ test_that_odin <- function(desc, code) {
   targets <- test_odin_targets()
   code_enq <- rlang::enquo(code)
   for (target in targets) {
+    opts <- list(odin.target = target,
+                 odin.rewrite_dims = target == "c")
     testthat::test_that(sprintf("%s (%s)", desc, target),
-                        withr::with_options(list(odin.target = target),
-                                            rlang::eval_tidy(code_enq)))
+                        withr::with_options(opts, rlang::eval_tidy(code_enq)))
   }
 }