+++ NEWS.md
@@ -1,3 +1,7 @@
+# odin 1.1.9
+
+* New option `rewrite_dims` (via `odin::odin_options`) which will attempt to simplify common dimensions. This can reduce the number of variables carried around in the model as these are typically very redundant and also known at compile time (mrc-2093)
+
 # odin 1.1.8
 
 * Annotate equations with `# ignore.unused` to locally suppress messages about unused variables (mrc-2122)
@@ -19,7 +23,7 @@
 * The basic infrastructure has been overhauled, which will make some alternative compilation targets easier to support. We now use `pkgbuild` for the compilation which should ease debugging, and odin code compiled into packages will no longer issue a slew of warnings (and cooperate with automatic routine registration). This refactor has caused a few minor breaking changes:
   - `$initial()` always requires time, even if it is ignored
   - `$set_user()` and construction no longer work with positional argument matching - all arguments must be named
-  - The `$ir` field has become a method; add parens after it
+  - The `$ir` field has become a method; add parentheses after it
   - The `compiler_warnings` option has been removed
 
 # odin 1.0.7